+++ packages/expo-updates-interface/CHANGELOG.md
@@ -11,6 +11,7 @@
 ### 💡 Others
 
 - Decouple from "bridge" in `expo-updates`. ([#27216](https://github.com/expo/expo/pull/27216) by [@kudo](https://github.com/kudo))
+- Migrated expo-updates-interface to Kotlin. ([#28033](https://github.com/expo/expo/pull/28033) by [@kudo](https://github.com/kudo))
 
 ## 0.15.3 - 2024-01-18